+++ Expanded Cheat Menu/Code/ReplacedFunctions.lua
@@ -809,8 +809,8 @@ function OnMsg.ClassesBuilt()
 		end
 
 		-- get them lazy drones working (bugfix for drones ignoring amounts less then their carry amount)
-		if UserSettings.DroneResourceCarryAmountFix then
-			ChoGGi.ComFuncs.FuckingDrones(self)
+		if UserSettings.DroneResourceCarryAmountFix and self:GetStoredAmount() > 1000 then
+			ChoGGi.ComFuncs.FuckingDrones(self,"single")
 		end
 
 		return ChoGGi_OrigFuncs.SingleResourceProducer_Produce(self, amount_to_produce, ...)