+++ Mods ChoGGi/Drones Carry Amount/Code/Script.lua
@@ -9,7 +9,9 @@ function OnMsg.ClassesBuilt()
 	local orig_SingleResourceProducer_Produce = SingleResourceProducer.Produce
 	function SingleResourceProducer:Produce(...)
 		-- get them lazy drones working
-		FuckingDrones(self)
+		if self:GetStoredAmount() > 1000 then
+			FuckingDrones(self,"single")
+		end
 		-- be on your way
 		return orig_SingleResourceProducer_Produce(self, ...)
 	end
@@ -20,18 +22,27 @@ function OnMsg.NewHour()
 	local labels = UICity.labels
 
 	-- Hey. Do I preach at you when you're lying stoned in the gutter? No!
-	local objs = labels.ResourceProducer or ""
-	for i = 1, #objs do
-		local obj = objs[i]
-		FuckingDrones(obj:GetProducerObj())
-		if obj.wasterock_producer then
-			FuckingDrones(obj.wasterock_producer)
+	local prods = labels.ResourceProducer or ""
+	for i = 1, #prods do
+		local prod = prods[i]
+		-- most are fine with GetProducerObj, but some like water extractor don't have one
+		local obj = prod:GetProducerObj() or prod
+		local func = obj.GetStoredAmount and "GetStoredAmount" or obj.GetAmountStored and "GetAmountStored"
+		if obj[func](obj) > 1000 then
+			FuckingDrones(obj)
+		end
+		obj = prod.wasterock_producer
+		if obj and obj:GetStoredAmount() > 1000 then
+			FuckingDrones(obj, "single")
 		end
 	end
 
-	objs = labels.BlackCubeStockpiles or ""
-	for i = 1, #objs do
-		FuckingDrones(objs[i])
+	prods = labels.BlackCubeStockpiles or ""
+	for i = 1, #prods do
+		local obj = prods[i]
+		if obj:GetStoredAmount() > 1000 then
+			FuckingDrones(obj)
+		end
 	end
 end